Fixes warning: implicit conversion loses integer precision: 'long long' to 'NSUInteger' (aka 'unsigned int') when built with -Wshorten-64-to-32

+++ AFNetworking/AFHTTPRequestOperation.m
@@ -331,7 +331,7 @@ didReceiveResponse:(NSURLResponse *)response
     if (self.outputStream) {
         [self.outputStream open];
     } else {
-        NSUInteger maxCapacity = MAX(llabs(response.expectedContentLength), kAFHTTPMinimumInitialDataCapacity);
+        NSUInteger maxCapacity = MAX((NSUInteger)llabs(response.expectedContentLength), kAFHTTPMinimumInitialDataCapacity);
         NSUInteger capacity = MIN(maxCapacity, kAFHTTPMaximumInitialDataCapacity);
         self.dataAccumulator = [NSMutableData dataWithCapacity:capacity];
     }
@@ -352,7 +352,7 @@ didReceiveResponse:(NSURLResponse *)response
     }
     
     if (self.downloadProgress) {
-        self.downloadProgress(self.totalBytesRead, self.response.expectedContentLength);
+        self.downloadProgress(self.totalBytesRead, (NSUInteger)self.response.expectedContentLength);
     }
 }
